Compile typescript after installing package
Anyone who wants to use this module in nodejs will need to compile the
typescript code, so I've copied the `npm run compile` target into the
`install` script.  This required adding typescript as a non-dev npm
dependency along with some other packages on the dev list.  I also
discovered that the `src` path resolution was working differently whether
I ran `npm install` on the package myself or whether the package was being
pulled through a `npm install` on a relying project.  Package-lock.json is
annoying, so I got rid of it.